extracted home page service and fixed error handling for certain api return cases
